working for admin and user

+++ src/main/java/com/devglan/service/UserService.java
@@ -8,12 +8,14 @@ import java.util.List;
 public interface UserService {
 
     User save(UserDto user);
+
     List<User> findAll();
-    void delete(int id);
+
+    void delete(Long id);
 
     User findOne(String username);
 
-    User findById(int id);
+    User findById(Long id);
 
     UserDto update(UserDto userDto);
 }
